clamped zoom values at ~17 digits of precision to prevent the fractal from being destroyed

+++ include/complex_coord.h
@@ -2,8 +2,8 @@
 #define COMPLEX_COORD_H
 
 typedef struct ComplexCoord {
-  double real;
-  double imaginary;
+  long double real;
+  long double imaginary;
 } c_coord_t;
 
 #endif
